cleanup of unused components and config
