+++ src/applications/_mock-form-ae-design-patterns/patterns/pattern1/TaskYellow/config/form.js
@@ -7,7 +7,6 @@ import manifest from '../../../../manifest.json';
 import content from '../../../../shared/locales/en/content.json';
 
 import IntroductionPage from '../../../../shared/components/pages/IntroductionPage1010ezr';
-import ConfirmationPage from '../../../../shared/components/pages/ConfirmationPage';
 import { VIEW_FIELD_SCHEMA } from '../../../../utils/constants';
 import { taskCompletePagePattern1 } from '../../../../shared/config/taskCompletePage';
 import { prefillTransformer } from '../../../../utils/helpers/prefill-transformer';
@@ -25,7 +24,6 @@ const formConfig = {
     Promise.resolve({ attributes: { confirmationNumber: '123123123' } }),
   trackingPrefix: 'task-yellow',
   introduction: IntroductionPage,
-  confirmation: ConfirmationPage,
   formId: VA_FORM_IDS.FORM_MOCK_AE_DESIGN_PATTERNS,
   getHelp: GetFormHelp,
   saveInProgress: {
