+++ src/applications/_mock-form-ae-design-patterns/patterns/pattern2/TaskBlue/config/veteranInfo.js
