--- src/applications/_mock-form-ae-design-patterns/patterns/pattern2/TaskBlue/content/contactInfo.js
@@ -1,7 +1,0 @@
-export const errorMessages = {
-  missingEmail: 'Add an email address to your profile',
-  missingHomePhone: 'Add a home phone number to your profile',
-  missingMobilePhone: 'Add a mobile phone number to your profile',
-  missingMailingAddress: 'Add a mailing address to your profile',
-  invalidMailingAddress: 'Add a valid mailing address to your profile',
-};
