+++ src/applications/_mock-form-ae-design-patterns/patterns/pattern2/TaskGray/form/config/form.js
@@ -1,33 +1,23 @@
 import React from 'react';
+
 import preSubmitInfo from 'platform/forms/preSubmitInfo';
 import FormFooter from 'platform/forms/components/FormFooter';
 import { VA_FORM_IDS } from 'platform/forms/constants';
 
+// application root imports
 import { scrollAndFocusTarget } from 'applications/_mock-form-ae-design-patterns/utils/focus';
+import { GetFormHelp } from 'applications/_mock-form-ae-design-patterns/shared/components/GetFormHelp';
+import { taskCompletePagePattern2 } from 'applications/_mock-form-ae-design-patterns/shared/config/taskCompletePage';
+import { VIEW_FIELD_SCHEMA } from 'applications/_mock-form-ae-design-patterns/utils/constants';
+
+// task specific imports
+import { ContactInformationInfoSection } from '../components/ContactInfo';
+import VeteranProfileInformation from '../components/VeteranProfileInformation';
 import IntroductionPage from '../containers/IntroductionPage';
-import ConfirmationPage from '../containers/ConfirmationPage';
-import { GetFormHelp } from '../components/GetFormHelp';
 import manifest from '../manifest.json';
 import { definitions } from './schemaImports';
 import profileContactInfo from './profileContactInfo';
-import { ContactInformationInfoSection } from '../components/ContactInfo';
-import VeteranProfileInformation from '../components/VeteranProfileInformation';
 import ReviewPage from '../../pages/ReviewPage';
-// chapter schema imports
-// import { applicantInformation } from './chapters/applicant';
-import { VIEW_FIELD_SCHEMA } from '../../../../../utils/constants';
-// import { taskCompletePage } from '../../../../../shared/config/taskCompletePage';
-
-import { taskCompletePagePattern2 } from '../../../../../shared/config/taskCompletePage';
-
-// import { serviceHistory } from './chapters/service';
-
-// import { loanScreener, loanHistory } from './chapters/loans';
-
-// import { fileUpload } from './chapters/documents';
-
-// TODO: When schema is migrated to vets-json-schema, remove common
-// definitions from form schema and get them from common definitions instead
 
 const formConfig = {
   rootUrl: manifest.rootUrl,
@@ -45,7 +35,6 @@ const formConfig = {
     reviewPageTitle: 'Review your request',
   },
   introduction: IntroductionPage,
-  confirmation: ConfirmationPage,
   formId: VA_FORM_IDS.FORM_26_1880,
   version: 0,
   prefillEnabled: true,
