+++ src/applications/_mock-form-ae-design-patterns/patterns/pattern2/TaskOrange/pages/introduction/IntroductionPage.jsx
@@ -1,11 +1,13 @@
 import PropTypes from 'prop-types';
 import React from 'react';
+import { connect } from 'react-redux';
+
 import { focusElement } from 'platform/utilities/ui';
 import FormTitle from 'platform/forms-system/src/js/components/FormTitle';
-import { connect } from 'react-redux';
 import { toggleValues } from 'platform/site-wide/feature-toggles/selectors';
 import FEATURE_FLAG_NAMES from 'platform/utilities/feature-toggles/featureFlagNames';
-import SaveInProgressInfo from '../components/SaveInProgressInfo';
+
+import SaveInProgressInfo from './SaveInProgressInfo';
 
 export class IntroductionPage extends React.Component {
   componentDidMount() {
