--- src/applications/_mock-form-ae-design-patterns/patterns/pattern2/containers/Pattern2.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export const Pattern2 = () => {
-  return <div>Pattern2</div>;
-};
